% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccs.R
\name{getccs}
\alias{getccs}
\title{Calculates clustering coefficients}
\usage{
getccs(net)
}
\arguments{
\item{net:}{an adjacency matrix}
}
\value{
a named vector with the local and global clustering coefficients
}
\description{
Calculates clustering coefficients
}

