% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lil_phi.R
\name{getlilphi}
\alias{getlilphi}
\title{Calculates little phi}
\usage{
getlilphi(data, net)
}
\arguments{
\item{data:}{a data.frame}

\item{net:}{an adjacency matrix}
}
\value{
a matrix, the estimate of \eqn{\phi}
}
\description{
Calculates little phi
}
\details{
WARNING: no checks on inputs, use with caution!
}

