% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_alt_rctr_know.R
\name{getpropaltrctrknow}
\alias{getpropaltrctrknow}
\title{Calculates the proportion that alter's recruiter (?) knows}
\usage{
getpropaltrctrknow(data, net)
}
\arguments{
\item{data:}{a data.frame}

\item{net:}{an adjacency matrix}
}
\value{
a matrix, little phi?
}
\description{
Calculates the proportion that alter's recruiter (?) knows
}

