% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HardimanKatzirCCs.R
\name{HardimanKatzirCCs}
\alias{HardimanKatzirCCs}
\title{Calculates HK clustering coefficients}
\usage{
HardimanKatzirCCs(recruiter, recruit, recruitee, deg, net)
}
\arguments{
\item{recruiter:}{a vector listing... recruiter ID number?}

\item{recruit:}{a vector listing... recruit ID number?}

\item{recruitee:}{a vector listing... recruitee ID number?}

\item{deg:}{a numeric vector of degree values for each recruit}

\item{net:}{the adjacency matrix for the network}
}
\value{
A named, numeric list containing the various clustering coefficients
}
\description{
Calculates HK clustering coefficients
}

